pyproject.toml
[tool.rye.workspace]
members = ["cli", "artagents"]

[project]
name = "eng-ai-agents"
version = "0.1.0"
description = "Course docker container"
authors = [
    { name = "Pantelis Monogioudis", email = "pantelis.monogioudis@aegean.ai" }
]
dependencies = [
    "pre-commit>=3.7.1",
    "jupyterlab-quarto>=0.3.3",
    "jupyter>=1.0.0",
    "torch>=2.3.1",
    "torchmetrics>=1.4.0.post0",
    "rich>=13.7.1",
    "lightning>=2.3.3",
    "hydra-core>=1.3.2",
    "rootutils>=1.0.7",
    "hydra-colorlog>=1.2.0",
    "hydra-optuna-sweeper>=1.2.0",
    "nats-py>=2.8.0",
    "requests>=2.32.3",
    "types-requests>=2.32.0.20240602",
    "grpcio>=1.64.1",
    "grpcio-tools<=1.64.1",
    "dapr>=1.13.0",
    "dapr-ext-grpc>=1.12.0",
    "s3fs<=2024.6.0",
    "datasets>=2.20.0",
    "pillow>=10.3.0",
    "dvc>=3.51.2",
    "dvclive>=3.46.1",
    "smart-open>=7.0.4",
    "geopy>=2.4.1",
    "typer>=0.12.5",
    "torchvision>=0.18.1",
    "torchinfo>=1.8.0",
    "torchviz>=0.0.2",
    "toml>=0.10.2",
    "s3path>=0.5.7",
    "tensorboard-logger>=0.1.0",
    "fastparquet>=2024.5.0",
    "numba>=0.60.0",
    "numexpr>=2.10.1",
    "bottleneck>=1.4.0",
    "ipykernel>=6.29.5",
    "tensorboard<2.17.0",
    "s3torchconnector>=1.2.3",
    "opencv-python-headless>=4.10.0.84",
    "scikit-learn>=1.5.1",
    "timm>=1.0.7",
    "python-dotenv>=1.0.1",
    "lightly>=1.5.10",
    "pinecone-client>=5.0.1",
    "umap-learn>=0.5.6",
    "pycocotools>=2.0.8",
    "torchsummary>=1.5.1",
    "seaborn>=0.13.2",
    "scikit-image>=0.24.0",
    "hist[plot]>=2.7.3",
    "tqdm>=4.66.5",
    "clearml>=1.16.4",
    "optuna<3.0.0",
    "ray>=2.35.0",
    "captum>=0.7.0",
    "bs4>=0.0.2",
    "pymongo>=4.10.1",
    "langchain>=0.3.10",
    "qdrant-client>=1.12.1",
    "langchain-community>=0.3.10",
    "langchain-openai>=0.2.11",
]
readme = "README.md"
requires-python = "== 3.11.9"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
     "pytest>=8.2.1",
    "pre-commit>=3.7.0",
    "ruff>=0.4.7",
    "pytest-check>=2.3.1",
    "pytest-cov>=5.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/aiml_common_solutions"]

[[tool.rye.sources]]
name = "pytorch"
url = "https://download.pytorch.org/whl/gpu"


[tool.pytest.ini_options]
addopts = "--cov"
filterwarnings = ["ignore::pytest.PytestCacheWarning"]
